refactor: isolate runtime waiting boundary
