+++ src/runtime/callback.rs
@@ -1,35 +1,9 @@
 use super::*;
 
 use crate::callbacks::{build_callback_url, generate_callback_token};
-use crate::types::WaitingReason;
+use crate::ingress::WakeHint;
 
 impl RuntimeHandle {
-    pub async fn latest_waiting_intents(&self) -> Result<Vec<WaitingIntentRecord>> {
-        let agent_id = self.agent_id().await?;
-        let mut records = self
-            .inner
-            .storage
-            .latest_waiting_intents()?
-            .into_iter()
-            .filter(|record| record.agent_id == agent_id)
-            .collect::<Vec<_>>();
-        records.sort_by(|left, right| right.created_at.cmp(&left.created_at));
-        Ok(records)
-    }
-
-    pub async fn latest_external_triggers(&self) -> Result<Vec<ExternalTriggerRecord>> {
-        let agent_id = self.agent_id().await?;
-        let mut records = self
-            .inner
-            .storage
-            .latest_external_triggers()?
-            .into_iter()
-            .filter(|record| record.target_agent_id == agent_id)
-            .collect::<Vec<_>>();
-        records.sort_by(|left, right| right.created_at.cmp(&left.created_at));
-        Ok(records)
-    }
-
     pub async fn create_callback(
         &self,
         summary: String,
@@ -132,81 +106,6 @@ impl RuntimeHandle {
             target_agent_id: self.agent_id().await?,
             scope,
             delivery_mode,
-        })
-    }
-
-    async fn current_waiting_work_item_anchor(&self, agent_id: &str) -> Result<Option<String>> {
-        let state = self.agent_state().await?;
-        let mut candidates = Vec::new();
-        candidates.extend(state.current_turn_work_item_id);
-        candidates.extend(state.current_work_item_id);
-        candidates.extend(
-            state
-                .working_memory
-                .current_working_memory
-                .current_work_item_id,
-        );
-        if let Some(current) = self.inner.storage.work_queue_prompt_projection()?.current {
-            candidates.push(current.id);
-        }
-
-        for candidate in candidates {
-            let Some(record) = self.inner.storage.latest_work_item(&candidate)? else {
-                continue;
-            };
-            if record.agent_id != agent_id || record.state == WorkItemState::Completed {
-                continue;
-            }
-            return Ok(Some(record.id));
-        }
-        Ok(None)
-    }
-
-    pub async fn cancel_waiting(&self, waiting_intent_id: &str) -> Result<CancelWaitingResult> {
-        let waiting = self
-            .latest_waiting_intents()
-            .await?
-            .into_iter()
-            .find(|record| record.id == waiting_intent_id)
-            .ok_or_else(|| anyhow!("waiting intent {} not found", waiting_intent_id))?;
-        let now = Utc::now();
-
-        let updated_waiting = if waiting.status == WaitingIntentStatus::Cancelled {
-            waiting.clone()
-        } else {
-            let mut updated = waiting.clone();
-            updated.status = WaitingIntentStatus::Cancelled;
-            updated.cancelled_at = Some(now);
-            self.inner.storage.append_waiting_intent(&updated)?;
-            updated
-        };
-
-        if let Some(descriptor) = self
-            .latest_external_triggers()
-            .await?
-            .into_iter()
-            .find(|record| record.external_trigger_id == waiting.external_trigger_id)
-        {
-            if descriptor.status != ExternalTriggerStatus::Revoked {
-                let mut revoked = descriptor;
-                revoked.status = ExternalTriggerStatus::Revoked;
-                revoked.revoked_at = Some(now);
-                self.inner.storage.append_external_trigger(&revoked)?;
-            }
-        }
-
-        self.inner.storage.append_event(&AuditEvent::new(
-            "waiting_intent_cancelled",
-            serde_json::json!({
-                "waiting_intent_id": updated_waiting.id,
-                "external_trigger_id": updated_waiting.external_trigger_id,
-            }),
-        ))?;
-
-        Ok(CancelWaitingResult {
-            waiting_intent_id: updated_waiting.id,
-            external_trigger_id: updated_waiting.external_trigger_id,
-            status: updated_waiting.status,
         })
     }
 
@@ -359,204 +258,6 @@ impl RuntimeHandle {
             delivery_mode: updated_descriptor.delivery_mode,
             disposition,
         })
-    }
-
-    pub(super) async fn active_waiting_intent_summaries(
-        &self,
-    ) -> Result<Vec<WaitingIntentSummary>> {
-        Ok(self
-            .latest_waiting_intents()
-            .await?
-            .into_iter()
-            .filter(|record| record.status == WaitingIntentStatus::Active)
-            .map(|record| WaitingIntentSummary {
-                id: record.id,
-                scope: record.scope,
-                description: record.description,
-                source: record.source,
-                resource: record.resource,
-                condition: record.condition,
-                delivery_mode: record.delivery_mode,
-                status: record.status,
-                trigger_count: record.trigger_count,
-                created_at: record.created_at,
-                cancelled_at: record.cancelled_at,
-                last_triggered_at: record.last_triggered_at,
-            })
-            .collect())
-    }
-
-    pub(super) async fn active_work_item_waiting_intent_count(&self) -> Result<usize> {
-        Ok(self
-            .latest_waiting_intents()
-            .await?
-            .into_iter()
-            .filter(|record| record.status == WaitingIntentStatus::Active)
-            .filter(|record| record.scope == ExternalTriggerScope::WorkItem)
-            .count())
-    }
-
-    pub(super) async fn active_external_trigger_summaries(
-        &self,
-    ) -> Result<Vec<ExternalTriggerSummary>> {
-        Ok(self
-            .latest_external_triggers()
-            .await?
-            .into_iter()
-            .filter(|record| record.status == ExternalTriggerStatus::Active)
-            .map(|record| ExternalTriggerSummary {
-                external_trigger_id: record.external_trigger_id,
-                target_agent_id: record.target_agent_id,
-                waiting_intent_id: record.waiting_intent_id,
-                scope: record.scope,
-                delivery_mode: record.delivery_mode,
-                status: record.status,
-                delivery_count: record.delivery_count,
-                created_at: record.created_at,
-                revoked_at: record.revoked_at,
-                last_delivered_at: record.last_delivered_at,
-            })
-            .collect())
-    }
-
-    pub(super) async fn reconcile_waiting_contract(
-        &self,
-        message: &MessageEnvelope,
-        pre_cleanup_closure: &ClosureDecision,
-    ) -> Result<()> {
-        let active_waiting = self
-            .latest_waiting_intents()
-            .await?
-            .into_iter()
-            .filter(|record| record.status == WaitingIntentStatus::Active)
-            .filter(|record| record.scope == ExternalTriggerScope::WorkItem)
-            .collect::<Vec<_>>();
-        if active_waiting.is_empty() {
-            return Ok(());
-        }
-
-        let current_work_item_id = self
-            .inner
-            .storage
-            .work_queue_prompt_projection()?
-            .current
-            .as_ref()
-            .map(|item| item.id.clone());
-        let prior_anchor_id = {
-            let guard = self.inner.agent.lock().await;
-            guard
-                .state
-                .working_memory
-                .current_working_memory
-                .current_work_item_id
-                .clone()
-        };
-
-        if current_work_item_id.is_none() {
-            let cancelled_ids = self
-                .cancel_waiting_intents(
-                    active_waiting
-                        .iter()
-                        .map(|record| record.id.clone())
-                        .collect(),
-                )
-                .await?;
-            self.inner.storage.append_event(&AuditEvent::new(
-                "missing_current_work_item_before_wait",
-                serde_json::json!({
-                    "agent_id": self.agent_id().await?,
-                    "message_id": message.id,
-                    "waiting_intent_ids": cancelled_ids,
-                    "prior_current_work_item_id": prior_anchor_id,
-                    "closure": pre_cleanup_closure,
-                }),
-            ))?;
-            return Ok(());
-        }
-
-        let mut stale_ids = Vec::new();
-        let anchor_switched = prior_anchor_id.is_some()
-            && prior_anchor_id.as_deref() != current_work_item_id.as_deref();
-        if anchor_switched {
-            stale_ids.extend(
-                active_waiting
-                    .iter()
-                    .filter(|record| record.created_at < message.created_at)
-                    .map(|record| record.id.clone()),
-            );
-        } else if pre_cleanup_closure.waiting_reason != Some(WaitingReason::AwaitingExternalChange)
-        {
-            stale_ids.extend(
-                active_waiting
-                    .iter()
-                    .filter(|record| record.created_at < message.created_at)
-                    .map(|record| record.id.clone()),
-            );
-        }
-
-        if stale_ids.is_empty() {
-            return Ok(());
-        }
-
-        let reason = if anchor_switched {
-            "active_work_switched"
-        } else {
-            "closure_no_longer_waiting_on_external_change"
-        };
-        let cancelled_ids = self.cancel_waiting_intents(stale_ids).await?;
-        self.inner.storage.append_event(&AuditEvent::new(
-            "stale_waiting_intents_cancelled",
-            serde_json::json!({
-                "agent_id": self.agent_id().await?,
-                "message_id": message.id,
-                "reason": reason,
-                "waiting_intent_ids": cancelled_ids,
-                "prior_current_work_item_id": prior_anchor_id,
-                "current_work_item_id": current_work_item_id,
-                "closure": pre_cleanup_closure,
-            }),
-        ))?;
-        Ok(())
-    }
-
-    pub(super) async fn cancel_work_item_waiting_intents(
-        &self,
-        work_item_id: &str,
-        reason: &'static str,
-    ) -> Result<Vec<String>> {
-        let waiting_intent_ids = self
-            .latest_waiting_intents()
-            .await?
-            .into_iter()
-            .filter(|record| record.status == WaitingIntentStatus::Active)
-            .filter(|record| record.scope == ExternalTriggerScope::WorkItem)
-            .filter(|record| record.work_item_id.as_deref() == Some(work_item_id))
-            .map(|record| record.id)
-            .collect::<Vec<_>>();
-        if waiting_intent_ids.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        let cancelled_ids = self.cancel_waiting_intents(waiting_intent_ids).await?;
-        self.inner.storage.append_event(&AuditEvent::new(
-            "work_item_waiting_intents_cancelled",
-            serde_json::json!({
-                "agent_id": self.agent_id().await?,
-                "work_item_id": work_item_id,
-                "reason": reason,
-                "waiting_intent_ids": cancelled_ids,
-            }),
-        ))?;
-        Ok(cancelled_ids)
-    }
-
-    async fn cancel_waiting_intents(&self, waiting_intent_ids: Vec<String>) -> Result<Vec<String>> {
-        let mut cancelled = Vec::new();
-        for waiting_intent_id in waiting_intent_ids {
-            self.cancel_waiting(&waiting_intent_id).await?;
-            cancelled.push(waiting_intent_id);
-        }
-        Ok(cancelled)
     }
 }
